# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coopvest Africa Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: getLoans
      tags: [loans]
      summary: List all loans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, defaulted, repaid, rejected]
        - name: memberId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Loans list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoansListResponse"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanBody"
      responses:
        "201":
          description: Created loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/approve:
    post:
      operationId: approveLoan
      tags: [loans]
      summary: Approve a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/reject:
    post:
      operationId: rejectLoan
      tags: [loans]
      summary: Reject a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectLoanBody"
      responses:
        "200":
          description: Loan rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
  /contributions:
    get:
      operationId: getContributions
      tags: [contributions]
      summary: List all contributions
      parameters:
        - name: memberId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Contributions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributionsListResponse"
    post:
      operationId: createContribution
      tags: [contributions]
      summary: Record a contribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContributionBody"
      responses:
        "201":
          description: Created contribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
...
  /investments:
    get:
      operationId: getInvestments
      tags: [investments]
      summary: List all investments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Investments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentsListResponse"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create an investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestmentBody"
      responses:
        "201":
          description: Created investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
...
components:
  schemas:
...
    Loan:
      type: object
      properties:
        id:
          type: integer
        loanId:
          type: string
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        balance:
          type: number
        interestRate:
          type: number
        tenure:
          type: integer
        status:
          type: string
          enum: [pending, active, defaulted, repaid, rejected]
        purpose:
          type: string
        disbursedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        monthlyPayment:
          type: number
        nextPaymentDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
      required: [id, loanId, memberId, memberName, amount, balance, interestRate, tenure, status, purpose, createdAt]
    LoansListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateLoanBody:
      type: object
      properties:
        memberId:
          type: integer
        amount:
          type: number
        tenure:
          type: integer
        purpose:
          type: string
        interestRate:
          type: number
      required: [memberId, amount, tenure, purpose]
    RejectLoanBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...
    Contribution:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        month:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
          enum: [paid, pending, overdue]
        transactionRef:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, memberId, memberName, amount, month, paymentMethod, status, createdAt]
    ContributionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contribution"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateContributionBody:
      type: object
      properties:
        memberId:
          type: integer
        amount:
          type: number
        month:
          type: string
        paymentMethod:
          type: string
      required: [memberId, amount, month, paymentMethod]
...
    Investment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        amount:
          type: number
        currentValue:
          type: number
        returns:
          type: number
        returnPercentage:
          type: number
        status:
          type: string
          enum: [active, matured, liquidated, pending]
        startDate:
          type: string
          format: date
        maturityDate:
          type: string
          format: date
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, amount, currentValue, returns, returnPercentage, status, startDate, createdAt]
    InvestmentsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Investment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateInvestmentBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        amount:
          type: number
        startDate:
          type: string
          format: date
        maturityDate:
          type: string
          format: date
        description:
          type: string
      required: [name, type, amount, startDate]
...