# lib/api-zod/src/generated/types/createLoanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coopvest Africa Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLoanBody {
  memberId: number;
  amount: number;
  tenure: number;
  purpose: string;
  interestRate?: number;
}