# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coopvest Africa Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentStatus } from "./investmentStatus";

export interface Investment {
  id: number;
  name: string;
  type: string;
  amount: number;
  currentValue: number;
  returns: number;
  returnPercentage: number;
  status: InvestmentStatus;
  startDate: Date;
  maturityDate?: Date;
  description?: string;
  createdAt: Date;
}