# lib/api-zod/src/generated/types/loansListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coopvest Africa Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Loan } from "./loan";

export interface LoansListResponse {
  data: Loan[];
  total: number;
  page: number;
  limit: number;
}